+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Key figures, contacts, links, and embed models to country and region pages.
+- Create endpoints for key figures and embeds that respond to private/public setting.
+
+### Changed
+
+- Altered the emergency key figure model to use a text type, ie "15%."
+
 ## 1.0.0
 
 ### Added